% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassify.R
\name{biomassify}
\alias{biomassify}
\title{Calcula el peso seco}
\usage{
biomassify(especie, long, k = 0.1)
}
\arguments{
\item{especie}{cadena}

\item{long}{longitud}

\item{k}{porcentage sin agua}
}
\value{
volumen
}
\description{
Calcula el peso seco
}
