% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoobiomass.R
\name{zoobiomass}
\alias{zoobiomass}
\title{Title}
\usage{
zoobiomass(especie, long)
}
\arguments{
\item{especie}{nombre}

\item{long}{longitu in micrometers}
}
\value{
volumen en mm cubicos
}
\description{
Title
}
\examples{
long <- 320 # in micrometers
zoobiomass("brac", long)
}
